solucoes/eduardoreisal/02-telegram-bot/requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
fonttools==4.39.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.1
hyperframe==6.0.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
snscrape==0.6.1.20230314
soupsieve==2.4
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
